+++ push_swap/strings.h
@@ -6,7 +6,7 @@
 /*   By: herbie <herbie@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/17 22:02:00 by herbie            #+#    #+#             */
-/*   Updated: 2023/02/21 20:37:11 by herbie           ###   ########.fr       */
+/*   Updated: 2023/02/22 10:35:48 by herbie           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@
 # include <stdlib.h>
 
 size_t	ft_strlen(const char *s);
-int		ft_atoi(const char *nptr);
+long	ft_atoi(const char *nptr);
 void	*ft_calloc(size_t nmemb, size_t size);
 void	*ft_memset(void *s, int c, size_t n);
 char	*ft_substr(char const *s, unsigned int start, size_t len);
